i have change and added the Pull class to First_pull class. i have created some functions and also added object properties.
